Simplify TMUX setup script
* The `tt` alias only requires one swap pane command. The other command
was present since the first one was ignored (probably because it started
executing before the previous command finished so it was ignored)
* Now things appear to work as they should. The idea here was to have
the first command start tmux with `tat`, and then the second will
split the window vertically (my preferred setup so that I can send rspec
test to the other pane) and swap panes since I prefer to work on the
left side of the screen. The same alias is used for both since `tat` is
idempotent and will not do anything the second time.